+++ packages/compartment-mapper/src/import.js
@@ -26,11 +26,12 @@ export const loadLocation = async (read, moduleLocation) => {
   );
 
   const execute = async (options = {}) => {
-    const { globals, modules, transforms } = options;
+    const { globals, globalLexicals, modules, transforms } = options;
     const makeImportHook = makeImportHookMaker(read, packageLocation);
     const compartment = assemble(compartmentMap, {
       makeImportHook,
       globals,
+      globalLexicals,
       modules,
       transforms
     });